[FE] REFACTOR: 반복되는 로직 utils 파일과 custom hook 으로 분리 #1649
