+++ frontend/src/Cabinet/components/Card/DisplayStyleCard/displayStyleInitializer.ts
@@ -3,23 +3,11 @@ import {
   DisplayStyleToggleType,
   DisplayStyleType,
 } from "@/Cabinet/types/enum/displayStyle.type.enum";
-
-export const getDisplayStyleFromLocalStorage: () => DisplayStyleToggleType =
-  () => {
-    return (
-      (localStorage.getItem(
-        "display-style-toggle"
-      ) as DisplayStyleToggleType) || DisplayStyleToggleType.DEVICE
-    );
-  };
-
-export const updateBodyDisplayStyle = (style: DisplayStyleType) => {
-  document.body.setAttribute("display-style", style);
-};
-
-export const isDeviceDarkMode = () => {
-  return window.matchMedia("(prefers-color-scheme: dark)");
-};
+import {
+  getDisplayStyleFromLocalStorage,
+  isDeviceDarkMode,
+  updateBodyDisplayStyle,
+} from "@/Cabinet/utils/displayStyleUtils";
 
 (function () {
   const isClient = typeof window !== "undefined";
